Large Commit (chams, doubletap, etc) Read Body
Added tickbase exploit with indicator
Added chams (using DrawModelExecute) (flat, texture, fresnel. With overlay chams)
Added debug flag for disabling friend related checks (temp fix for the Steam ID bug)
Thirdperson now collides with map geometry
Pickup respawn timers now center to the pickup and not the pickuper
Added injection script (debug.sh) that attaches GDB and makes it interactable
Set build mode to force C++23 standards
Added libCURL as a dependency (TODO: add to package scripts)
Added support for the PiShock API (debug)
Added Watermark toggle
Added custom Watermark text
More undocumented changes regarding hooked methods, dependencies, README changes, Makefile changes, build.sh changes, and general code changes

+++ build.sh
@@ -1,23 +1,17 @@
 #!/bin/bash
 
-# Clone and build funchook
-if [ -d "libs/funchook" ]; then
+# Clone and build funchook, if not done already.
+if [ -d "libs/funchook" ] && [ ! -f "libs/funchook/build/libfunchook.a" ] && [ ! -f "libs/funchook/build/libdistorm.a" ]; then
     rm -rf libs
 fi
 
 if [ ! -d "libs/funchook" ]; then
     mkdir libs
-    cd libs
-    git clone https://github.com/Doctor-Coomer/funchook.git
-    cd funchook
-    mkdir build
-    cd build
-    cmake -DCMAKE_BUILD_TYPE=Release ..
+    git clone https://github.com/Doctor-Coomer/funchook.git libs/
+    mkdir libs/funchook/build
+    cmake -DCMAKE_BUILD_TYPE=Release libs/funchook/build
     make
 fi
-
-# Escape back to the root project directory
-cd ../../..
 
 # Build hack
 make clean && make
